requirements.txt
absl-py==1.4.0
amqp==5.1.1
astunparse==1.6.3
attrs==23.1.0
billiard==3.6.4.0
blinker==1.6.2
Bottleneck==1.3.7
cachetools==5.3.0
category-encoders==1.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cornac==1.15.1
cycler==0.11.0
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
flatbuffers==23.3.3
fonttools==4.39.3
fsspec==2023.4.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.14.1
hypothesis==6.75.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.0.3
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
kombu==5.2.4
libclang==16.0.0
lightfm==1.17
lightgbm==3.3.5
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
memory-profiler==0.61.0
ml-dtypes==0.1.0
mpmath==1.3.0
multimethod==1.9.1
mypy-extensions==1.0.0
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
pandera==0.14.5
patsy==0.5.3
Pillow==9.5.0
powerlaw==1.5
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
recommenders==1.1.1
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
scikit-learn==1.0.2
scikit-surprise==1.1.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sortedcontainers==2.4.0
statsmodels==0.13.5
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.2
wrapt==1.14.1
zipp==3.15.0
